+++ core/src/com/template/game/screen/ScreenInterface.java
@@ -9,9 +9,9 @@ import com.badlogic.gdx.scenes.scene2d.Stage;
 
 public interface ScreenInterface extends Screen {
 
-    void draw(Batch batch, OrthographicCamera cam, float delta);
+    void draw(Batch batch, OrthographicCamera cam, double delta);
 
-    void draw(ShapeRenderer shapeRenderer, OrthographicCamera cam, float delta);
+    void draw(ShapeRenderer shapeRenderer, OrthographicCamera cam, double delta);
 
     void setupUI(Stage stage);
